requirements.txt
aiogram
python-dotenv
Pillow
pydub
ffmpeg-python
aiosqlite
diffusers
torch
accelerate
transformers
openai-whisper